+++ __mocks__/figma/sample_dump.json
@@ -1,0 +1,1 @@
+undefined
