+++ __mocks__/figma/sample_output.styledictionary.json
@@ -1,0 +1,224 @@
+{
+  "color": {
+    "base": {
+      "default": {
+        "value": "#5FC2E6",
+        "comment": "seagull"
+      },
+      "id1": {
+        "value": "#5EC1D1",
+        "comment": "viking"
+      },
+      "id2": {
+        "value": "#77C8D6",
+        "comment": "viking"
+      },
+      "id3": {
+        "value": "#8DCFDB",
+        "comment": "halfBaked"
+      },
+      "id4": {
+        "value": "#A0D6E0",
+        "comment": "powderBlue"
+      },
+      "id5": {
+        "value": "#B3DDE6",
+        "comment": "powderBlue"
+      },
+      "id6": {
+        "value": "#D4EBF0",
+        "comment": "pattensBlue"
+      },
+      "id7": {
+        "value": "#E5F3FA",
+        "comment": "solitude"
+      },
+      "id8": {
+        "value": "#F2F9FD",
+        "comment": "aliceBlue"
+      },
+      "id9": {
+        "value": "#227D3E",
+        "comment": "salem"
+      },
+      "id10": {
+        "value": "#F55E5E",
+        "comment": "carnation"
+      },
+      "id11": {
+        "value": "#3E874A",
+        "comment": "seaGreen"
+      },
+      "id12": {
+        "value": "#559158",
+        "comment": "hippieGreen"
+      },
+      "id13": {
+        "value": "#6B9D6A",
+        "comment": "aquaForest"
+      },
+      "id14": {
+        "value": "#81A97C",
+        "comment": "amulet"
+      },
+      "id15": {
+        "value": "#97B690",
+        "comment": "norway"
+      },
+      "id16": {
+        "value": "#C0D2BA",
+        "comment": "paleLeaf"
+      },
+      "id17": {
+        "value": "#D4E0D0",
+        "comment": "tasman"
+      },
+      "id18": {
+        "value": "#E9EFE7",
+        "comment": "grayNurse"
+      }
+    },
+    "border": {
+      "inside": {
+        "value": "#FF9696",
+        "comment": "geraldine"
+      },
+      "outside": {
+        "value": "#FF9696",
+        "comment": "geraldine"
+      },
+      "center": {
+        "value": "#FF9696",
+        "comment": "geraldine"
+      }
+    },
+    "shadow": {
+      "simple": {
+        "dropShadow": {
+          "value": "#000000",
+          "comment": "black"
+        },
+        "innerShadow": {
+          "value": "#000000",
+          "comment": "black"
+        }
+      }
+    }
+  },
+  "font": {
+    "typo": {
+      "heading1": {
+        "value": "firaSansBold"
+      },
+      "heading2": {
+        "value": "firaSansRegular"
+      },
+      "paragraph": {
+        "value": "firaSansBook"
+      },
+      "quote": {
+        "value": "helveticaBold"
+      }
+    }
+  },
+  "lineHeight": {
+    "font": {
+      "heading1": {
+        "value": 57.6
+      },
+      "heading2": {
+        "value": 42.2
+      },
+      "paragraph": {
+        "value": 21.1
+      },
+      "quote": {
+        "value": 16.4
+      }
+    }
+  },
+  "size": {
+    "font": {
+      "heading1": {
+        "value": 48
+      },
+      "heading2": {
+        "value": 36
+      },
+      "paragraph": {
+        "value": 18
+      },
+      "quote": {
+        "value": 14
+      }
+    },
+    "border": {
+      "inside": {
+        "value": "1"
+      },
+      "outside": {
+        "value": "1"
+      },
+      "gradientWithCorners": {
+        "value": "1"
+      },
+      "center": {
+        "value": "1"
+      }
+    }
+  },
+  "weight": {
+    "fontFamily": {
+      "firaSansBold": {
+        "value": 700
+      },
+      "firaSansBook": {
+        "value": 350
+      },
+      "firaSansRegular": {
+        "value": 400
+      },
+      "helveticaBold": {
+        "value": 700
+      }
+    },
+    "font": {
+      "heading1": {
+        "value": 700
+      },
+      "heading2": {
+        "value": 400
+      },
+      "paragraph": {
+        "value": 350
+      },
+      "quote": {
+        "value": 700
+      }
+    }
+  },
+  "asset": {
+    "font": {
+      "firaSansBook": {
+        "name": {
+          "value": "FiraSans-Book"
+        }
+      },
+      "helveticaBold": {
+        "name": {
+          "value": "Helvetica-Bold"
+        }
+      },
+      "firaSansRegular": {
+        "name": {
+          "value": "FiraSans-Regular"
+        }
+      },
+      "firaSansBold": {
+        "name": {
+          "value": "FiraSans-Bold"
+        }
+      }
+    }
+  }
+}
