+++ __mocks__/sketch/sample_output.document.json
@@ -557,6 +557,5 @@
     "Helvetica-Bold",
     "FiraSans-Regular",
     "FiraSans-Book"
-  ],
-  "sketchVersion": "55"
+  ]
 }
