+++ __mocks__/sketch/sample_output.styledictionary.json
