+++ lib/mappers/__tests__/__fixtures__/mapper.figma.js
@@ -1,4 +1,4 @@
-const sampleOutput = require('../../../../__mocks__/sample_dump_figma.json');
+const sampleOutput = require('../../../../__mocks__/figma/sample_dump.json');
 const { figma2jsonFromJSON, findIds } = require('../../../figma');
 const { uniqueArray } = require('../../../utils');
